+++ src/mz_os_win32.h
@@ -32,9 +32,9 @@ typedef void* DIR;
 /***************************************************************************/
 
 int32_t mz_win32_rand(uint8_t *buf, int32_t size);
-int16_t mz_win32_get_file_date(const char *path, uint32_t *dos_date);
-int16_t mz_win32_set_file_date(const char *path, uint32_t dos_date);
-int16_t mz_win32_make_dir(const char *path);
+int32_t mz_win32_get_file_date(const char *path, uint32_t *dos_date);
+int32_t mz_win32_set_file_date(const char *path, uint32_t dos_date);
+int32_t mz_win32_make_dir(const char *path);
 DIR*    mz_win32_open_dir(const char *path);
 struct 
 dirent* mz_win32_read_dir(DIR *dir);
